Clean code, fix typo on output specification, change int converted types
