+++ Testing/Expected/getAbiDeclaration.ts
@@ -2,7 +2,6 @@ export const getAbiDecTest0 = `export interface ITest0{
 }
 export interface ITest0Connected {
 address: string
-balance: ABIFuncParamlessCallConnected
 }
 export type ITest0Constructor = ABIFuncParamlessSendConnected
 `
@@ -11,37 +10,34 @@ export const getAbiDecTest1 = `export interface ITest1{
 }
 export interface ITest1Connected {
 address: string
-balance: ABIFuncParamlessCallConnected
 }
 export type ITest1Constructor = ABIFuncParamlessSendConnected
 `
 
 export const getAbiDecTest2 = `export interface ITest2{
-function1: ABIFuncSend<{a: string}>
-function2: ABIFuncSend<{a: string,b: BN}>
-function3: ABIFuncSend<{a: string,b: BN}>
+function1: ABIFuncSend<{a: string | Buffer}>
+function2: ABIFuncSend<{a: string | Buffer,b: BN | Buffer}>
+function3: ABIFuncSend<{a: string | Buffer,b: BN | Buffer}>
 function0: ABIFuncParamlessSend
 }
 export interface ITest2Connected {
 address: string
-balance: ABIFuncParamlessCallConnected
-function1: ABIFuncSendConnected<{a: string}>;
-function2: ABIFuncSendConnected<{a: string,b: BN}>;
-function3: ABIFuncSendConnected<{a: string,b: BN}>;
+function1: ABIFuncSendConnected<{a: string | Buffer}>;
+function2: ABIFuncSendConnected<{a: string | Buffer,b: BN | Buffer}>;
+function3: ABIFuncSendConnected<{a: string | Buffer,b: BN | Buffer}>;
 function0: ABIFuncParamlessSendConnected;
 }
 export type ITest2Constructor = ABIFuncParamlessSendConnected
 `
 
 export const getAbiDecTest3 = `export interface ITest3{
-overloaded: ABIFuncSend<{a: string}> | ABIFuncSend<{a: BN}>
-Test1: ABIFuncParamlessSend
+overloaded: ABIFuncSend<{a: string | Buffer}> | ABIFuncSend<{a: BN | Buffer}>
+Test3: ABIFuncParamlessSend
 }
 export interface ITest3Connected {
 address: string
-balance: ABIFuncParamlessCallConnected
-overloaded: ABIFuncSendConnected<{a: string}> | ABIFuncSendConnected<{a: BN}>
-Test1: ABIFuncParamlessSendConnected;
+overloaded: ABIFuncSendConnected<{a: string | Buffer}> | ABIFuncSendConnected<{a: BN | Buffer}>
+Test3: ABIFuncParamlessSendConnected;
 }
 export type ITest3Constructor = never
 `
