+++ Testing/__tests__/getAbiDeclaration.spec.ts
@@ -1,5 +1,12 @@
-import { getAbiDeclaration } from '../../lib/src/builder/getAbiDeclaration';
-import { getAbiDecTest0, getAbiDecTest1, getAbiDecTest2 , getAbiDecTest3} from '../Expected';
+import { getAbiDeclaration, parseAbi } from '../../lib/src/builder';
+import { getAbiDecTest0, 
+				 getAbiDecTest1, 
+				 getAbiDecTest2,
+				 getAbiDecTest3, 
+				 parseAbiTest0, 
+				 parseAbiTest1, 
+				 parseAbiTest2, 
+				 parseAbiTest3 } from '../Expected';
 
 const test0File = require('../abis/test0.json')
 const test1File = require('../abis/test1.json')
