+++ lib/src/builder/index.ts
@@ -1,2 +1,3 @@
 export { interfaces, imports } from './staticContent';
 export { getAbiDeclaration } from './getAbiDeclaration'
+export { parseAbi } from './parseAbi'
