+++ lib/src/builder/staticContent.ts
@@ -47,5 +47,4 @@ interface ICallTxObj {
 }`;
 
 export const defaultProperties = `address: string
-balance: ABIFuncParamlessCallConnected
 `
