+++ lib/src/cli/options.ts
@@ -1,6 +1,6 @@
 export const optionDefinitions = [
     { name: 'files', alias: 'f', type: String, multiple: true },
     { name: 'outputConfigurations', alias: 'c', type: String, multiple: true},
-    { name: 'outputFiles', alias: 'o', type: String },
+    { name: 'outputFile', alias: 'o', type: String },
     { name: 'help', alias: 'h' }
 ]
