+++ package.json
@@ -41,7 +41,7 @@
     "transform": {
       "\\.(ts|tsx)$": "<rootDir>/node_modules/ts-jest/preprocessor.js"
     },
-    "testRegex": "/__tests__/.*\\.(ts|tsx|js)$"
+    "testRegex": "/__tests__/.*\\.(ts|tsx)$"
   },
   "bin": {
     "typ3-cli": "./dist/index.js"
